+++ vignette.h
@@ -52,8 +52,8 @@ public:
   int shapes_ys[max_shapes];
 
   float shapes_symb_output[max_shapes * nb_symbolic_outputs];
-  uint shape_is_bordering[max_shapes * max_shapes];
-  uint shape_is_containing[max_shapes * max_shapes];
+  float shape_is_bordering[max_shapes * max_shapes];
+  float shape_is_containing[max_shapes * max_shapes];
 
   void clear();
